+++ node/src/types/kernel/api/network/forwardProxy/response.d.ts
@@ -15,22 +15,47 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+/**
+ * Forward proxy
+ */
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    data: IData;
+    /**
+     * status message
+     */
+    msg: string;
+}
 
-/* directories */
-export * as asset from "./asset";
-export * as attr from "./attr";
-export * as block from "./block";
-export * as convert from "./convert";
-export * as export from "./export";
-export * as file from "./file";
-export * as filetree from "./filetree";
-export * as network from "./network";
-export * as notebook from "./notebook";
-export * as notification from "./notification";
-export * as query from "./query";
-export * as search from "./search";
-export * as storage from "./storage";
-export * as system from "./system";
-export * as template from "./template";
-
-/* files */
+/**
+ * Response information
+ */
+export interface IData {
+    /**
+     * response body
+     */
+    body: string;
+    /**
+     * response content-type
+     */
+    contentType: string;
+    /**
+     * response elapsed
+     */
+    elapsed: number;
+    /**
+     * response headers
+     */
+    headers: { [key: string]: string[] };
+    /**
+     * HTTP status code
+     */
+    status: number;
+    /**
+     * URL to request
+     */
+    url: string;
+}